+++ code/src/sixsq/nuvla/server/middleware/authn_info.clj
@@ -1,7 +1,8 @@
 (ns sixsq.nuvla.server.middleware.authn-info
   (:require
     [clojure.string :as str]
-    [sixsq.nuvla.auth.cookies :as cookies]))
+    [sixsq.nuvla.auth.cookies :as cookies]
+    [clojure.set :as set]))
 
 
 ;; NOTE: ring uses lower-cased values of header names!
@@ -34,23 +35,23 @@
     (let [user-id (first terms)
           claims  (seq (rest terms))
           session (first (keep is-session? (rest terms)))]
-      (cond-> {}
+      (cond-> {:claims #{user-id "group/nuvla-anon"}}
               user-id (assoc :user-id user-id)
-              claims (assoc :claims (set claims))
+              claims (update :claims set/union (set claims))
               session (assoc :session session)))))
 
 
 (defn cookie-info->authn-info
   "Returns a tuple with the user-id and list of claims based on the
    provided cookie info map."
-  [{:keys [user-id claims session]}]
+  [{:keys [user-id active-claim claims session]}]
   (when user-id
     (let [claims (set (remove str/blank? (-> claims
                                              (or "")
                                              (str/split #"\s+")
                                              (conj session))))]
       (cond-> {}
-              user-id (assoc :user-id user-id)
+              user-id (assoc :user-id (or active-claim user-id))
               claims (assoc :claims (set claims))
               session (assoc :session session)))))
 
@@ -63,10 +64,9 @@
           (cookie-info->authn-info)))
 
 
-(defn add-anon-role-and-user [{:keys [nuvla/authn] :as request}]
-  (let [{:keys [user-id claims]} authn
-        claims-updated (cond-> (conj (set claims) "group/nuvla-anon")
-                               user-id (conj user-id))]
+(defn add-anon-role [{:keys [nuvla/authn] :as request}]
+  (let [{:keys [claims]} authn
+        claims-updated (conj (set claims) "group/nuvla-anon")]
     (assoc-in request [:nuvla/authn :claims] claims-updated)))
 
 
@@ -86,5 +86,5 @@
   (fn [request]
     (-> request
         add-authn-info
-        add-anon-role-and-user
+        add-anon-role
         handler)))